# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateCurriculumRequest:
    properties:
      courses:
        type: string
      driver_license:
        type: string
      educations:
        items:
          $ref: '#/definitions/dto.CreateEducationRequest'
        type: array
      email:
        type: string
      full_name:
        maxLength: 50
        minLength: 5
        type: string
      image_url:
        type: string
      intro:
        type: string
      languages:
        type: string
      phone:
        type: string
      skills:
        type: string
      social_links:
        type: string
      user_id:
        type: string
      works:
        items:
          $ref: '#/definitions/dto.CreateWorkRequest'
        type: array
    required:
    - driver_license
    - email
    - full_name
    - intro
    - languages
    - phone
    - skills
    - user_id
    type: object
  dto.CreateEducationRequest:
    properties:
      degree:
        maxLength: 255
        minLength: 2
        type: string
      description:
        maxLength: 255
        minLength: 2
        type: string
      end_date:
        type: string
      institution:
        maxLength: 255
        minLength: 2
        type: string
      start_date:
        type: string
    required:
    - degree
    - description
    - institution
    - start_date
    type: object
...
  dto.CreateWorkRequest:
    properties:
      company:
        maxLength: 255
        minLength: 2
        type: string
      description:
        type: string
      end_date:
        description: Nullable para trabalhos atuais
        type: string
      position:
        maxLength: 255
        minLength: 2
        type: string
      start_date:
        type: string
    required:
    - company
    - position
    - start_date
    type: object
  dto.CurriculumBodyResponse:
    properties:
      content:
        type: string
    type: object
  dto.CurriculumListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CurriculumResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.CursorPagination'
    type: object
  dto.CurriculumResponse:
    properties:
      courses:
        type: string
      created_at:
        type: string
      driver_license:
        type: string
      educations:
        items:
          $ref: '#/definitions/dto.EducationResponse'
        type: array
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      image_url:
        type: string
      intro:
        type: string
      languages:
        type: string
      phone:
        type: string
      skills:
        type: string
      social_links:
        type: string
      updated_at:
        type: string
      works:
        items:
          $ref: '#/definitions/dto.WorkResponse'
        type: array
    type: object
...
  dto.CursorPagination:
    properties:
      cursor:
        type: string
      has_next_page:
        type: boolean
      limit:
        type: integer
      next_cursor:
        type: string
    type: object
...
  dto.EducationResponse:
    properties:
      created_at:
        type: string
      degree:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      institution:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: user not found
        type: string
    type: object
  dto.ErrorResponseServer:
    properties:
      error:
        example: Internal server error
        type: string
    type: object
  dto.ErrorResponseValidation:
    properties:
      error:
        example: Validation error
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        example: User deleted successfully
        type: string
    type: object
...
  dto.WorkResponse:
    properties:
      company:
        type: string
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      position:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: dafondeveloper@gmail.com
    name: DafonCV Support
    url: https://www.dafoncv.online
  description: API for consultancy business and cv generation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Dafon CV API
  version: "1.0"
paths:
...
  /api/v1/curriculums:
    post:
      consumes:
      - application/json
      description: Creates a new curriculum for a user
      parameters:
      - description: Create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCurriculumRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CurriculumResponse'
        "400":
          description: Validation error or invalid user ID
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Create curriculum
      tags:
      - curriculum
  /api/v1/curriculums/{curriculum_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a curriculum by ID
      parameters:
      - description: Curriculum ID
        in: path
        name: curriculum_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid curriculum ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: Curriculum not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Delete curriculum by ID
      tags:
      - curriculum
    get:
      consumes:
      - application/json
      description: Returns a curriculum by ID
      parameters:
      - description: Curriculum ID
        in: path
        name: curriculum_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CurriculumResponse'
        "400":
          description: Invalid curriculum ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: Curriculum not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Get curriculum by ID
      tags:
      - curriculum
  /api/v1/curriculums/get-all-by-user/{user_id}:
    get:
      consumes:
      - application/json
      description: Returns paginated list of curriculums for the given user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Cursor (UUID) to fetch items after
        in: query
        name: cursor
        type: string
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CurriculumListResponse'
        "400":
          description: Invalid user ID or pagination params
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Get all curriculums by user ID
      tags:
      - curriculum
  /api/v1/curriculums/get-body/{curriculum_id}:
    get:
      consumes:
      - application/json
      description: Returns the curriculum content in text format
      parameters:
      - description: Curriculum ID
        in: path
        name: curriculum_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CurriculumBodyResponse'
        "400":
          description: Invalid curriculum ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: Curriculum not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Get curriculum body
      tags:
      - curriculum
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  UserIDHeader:
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"